add basic shop

+++ RtvtPayment/app/src/main/java/sk/marcel/rtvt_payment/Item.kt
@@ -1,3 +1,3 @@
 package sk.marcel.rtvt_payment
 
-class Item(val index: Int, val name: String, val price: Int)
+class Item(val index: Int, val name: String, val price: Long)
